test fastPledgeCommitment
